+++ spring-api/src/main/java/eu/royalblackwater/api/fleet/FleetRepository.java
@@ -3,6 +3,6 @@ package eu.royalblackwater.api.fleet;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FleetRepository extends JpaRepository<FleetEntity, Long> {
+public interface FleetRepository extends JpaRepository<FleetEntity, Integer> {
     Optional<FleetEntity> findFirstByActiveTrueOrderBySortOrderAscIdAsc();
 }
