+++ spring-api/src/main/java/eu/royalblackwater/api/privacy/CookieConsentRepository.java
@@ -3,6 +3,6 @@ package eu.royalblackwater.api.privacy;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CookieConsentRepository extends JpaRepository<CookieConsentEntity, Long> {
+public interface CookieConsentRepository extends JpaRepository<CookieConsentEntity, Integer> {
     Optional<CookieConsentEntity> findFirstByConsentKeyOrderByCreatedAtDescIdDesc(String consentKey);
 }
